Add possibility to configure where a new page shall be inserted
This page allows the developer to decide, where a new page should
be inserted, when a new page is created. A new configuration option
called insert_page_at_beginning was added. The default is false.
If set to true, the new page will be inserted at the beginning
of a partial tree.

Technically the new page will have position 0 and all other
pages are updated with page.position = page.position + 1